Lau 2 with heat density map
- add layer lau2
- add heat density map LAU2 precomputed layer
- edit stats/selection/area endpoint to support heat_density_map_lau2

+++ app/main_api/models/heat_load_profile.py
@@ -28,8 +28,8 @@ class HeatLoadProfileNuts(db.Model):
     process = db.Column(db.String())
     unit = db.Column(db.String())
     value = db.Column(db.Numeric(precision=30, scale=10))
-    fk_nuts_gid = db.Column(db.Integer)
-    fk_time_id = db.Column(db.Integer)
+    fk_nuts_gid = db.Column(db.BigInteger)
+    fk_time_id = db.Column(db.BigInteger)
 
     nuts = db.relationship("Nuts")
     time = db.relationship("Time")
